requirements.txt
numpy==2.3.2
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
python-dotenv==1.0.1